\name{ABCExtremes-package}
\alias{ABCExtremes-package}
\alias{ABCExtremes}
\docType{package}
\title{ABC Extremes}
\description{
This package contains code to 1.) construct the summary statistic needed to facilitate approximate Bayesian 
computing for spatial extremes, and then 2.) run the ABC-Rejection method to produce a set of draws 
from the ABC posterior.  The intention is for the user to directly modify the command >abc.rej() 
to suit his or her purposes.  In general ABC fitting of max-stable processes is very computationally expensive, and requires
advanced programming skills with research computing.}
\details{
\tabular{ll}{
Package: \tab ABCExtremes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-14\cr
License: \tab What license is it under?\cr}
}
\author{Rob Erhardt
Maintainer: <erhardrj@wfu.edu>}
\references{
Erhardt, R., Smith, R. (2012). Approximate Bayesian Computing for  Spatial Extremes. Computational Statistics and Data Analysis Vol. 56:6, 1468-1481.}
\seealso{
See also: R packages SpatialExtremes, RandomFields}

